update deps and move tests into own file
